# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "GamiBook API",
        "description": "ESMAPP API - GamiBook é uma aplicação para a editora LIDEL que incentiva a aprendizagem através de exercícios gamificados. O token não expira já que a aplicação é destinada para o uso mobile.",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
        "parameters": {
...
            "bookCategory": {
                "name": "category",
                "in": "query",
                "description": "Identificador da categoria",
                "required": false,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
            "idBook": {
                "name": "idBook",
                "in": "path",
                "description": "Identificador do livro",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
            "idModule": {
                "name": "idModule",
                "in": "path",
                "description": "Identificador do módulo",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                }
            },
...
        },
...
    },
...
    "paths": {
...
        "/books": {
            "get": {
                "summary": "Lista de livros",
                "description": "Devolve uma lista de livros",
                "operationId": "findAllBooks",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/bookCategory"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/books/200/findAllBooks"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idCategory"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    }
                }
            },
            "post": {
                "summary": "Criar livros",
                "description": "Adiciona um livro à base de dados da aplicação",
                "operationId": "createBook",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/books/createBook"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/books/201/createBook"
                    },
                    "400": {
                        "$ref": "#/components/responses/books/400/createBook"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "406": {
                        "$ref": "#/components/responses/books/406/name_code"
                    }
                }
            }
        },
        "/books/{idBook}": {
            "get": {
                "summary": "Procurar um livro",
                "description": "Devolve a informação de um livro pelo id",
                "operationId": "findOneBook",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/books/200/findOneBook"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idBook"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idBook"
                    }
                }
            },
            "patch": {
                "summary": "Atualizar o livro",
                "description": "Atualiza alguns campos do livro",
                "operationId": "updateBook",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/books/updateBook"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/books/200/updateBook"
                    },
                    "400": {
                        "$ref": "#/components/responses/books/400/updateBook"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idBook"
                    }
                }
            },
            "delete": {
                "summary": "Apagar livros",
                "description": "Apaga um livro pelo id",
                "operationId": "deleteBook",
                "tags": [
                    "Livro",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/books/200/deleteBook"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idBook"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idBook"
                    }
                }
            }
        },
        "/books/{idBook}/modules": {
            "get": {
                "summary": "Lista de módulos de um livro",
                "description": "Devolve uma lista de módulos de acordo com o livro",
                "operationId": "findAllBookModules",
                "tags": [
                    "Módulo",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/modules/200/findAllBookModules"
                    },
                    "400": {
                        "$ref": "#/components/responses/general/400/idBook"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idBook"
                    }
                }
            },
            "post": {
                "summary": "Criar módulos",
                "description": "Adiciona um módulo à base de dados da aplicação",
                "operationId": "createBookModule",
                "tags": [
                    "Módulo",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/modules/moduleName"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/modules/201/createBookModule"
                    },
                    "400": {
                        "$ref": "#/components/responses/modules/400/createBookModule"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/idBook"
                    }
                }
            }
        },
        "/books/{idBook}/modules/{idModule}": {
            "patch": {
                "summary": "Atualizar o módulo",
                "description": "Atualiza alguns campos do módulo",
                "operationId": "updateBookModule",
                "tags": [
                    "Módulo",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    },
                    {
                        "$ref": "#/components/parameters/idModule"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/modules/moduleName"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/modules/200/updateBookModule"
                    },
                    "400": {
                        "$ref": "#/components/responses/modules/400/updateBookModule"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/book_module"
                    }
                }
            },
            "delete": {
                "summary": "Apagar módulos",
                "description": "Apaga um módulo pelo id",
                "operationId": "deleteBookModule",
                "tags": [
                    "Módulo",
                    "Rotas Privadas - Utilizador Administrador"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idBook"
                    },
                    {
                        "$ref": "#/components/parameters/idModule"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/modules/200/deleteBookModule"
                    },
                    "400": {
                        "$ref": "#/components/responses/modules/400/deleteBookModule"
                    },
                    "401": {
                        "$ref": "#/components/responses/general/401/unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/general/403/forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/general/404/book_module"
                    }
                }
            }
        }
    }
}